refactor(chat): move non-tool part predicates into tool-part-parsing
Fix round 1 on the tool-replay-reconciliation extraction: replace the
private duplicated isTextPart/isProviderVisibleReasoningPart/getFilePart
in tool-replay-reconciliation.ts (which had already drifted from
conversation.ts's isTextPart) with a real move of TextPartLike,
ReasoningPartLike, isTextPart, isReasoningPart,
isProviderVisibleReasoningPart, and getFilePart into
tool-part-parsing.ts. conversation.ts re-exports the public predicate
surface (isReasoningPart, isTextPart) so external callers are
unaffected.
